chore(cubesql): Initial support for custom protocols (#8572)
