+++ rust/cubesql/cubesql/src/compile/engine/context_postgresql.rs
@@ -1,38 +1,6 @@
-use std::{any::Any, collections::HashMap, sync::Arc};
+use std::sync::Arc;
 
-use async_trait::async_trait;
-use cubeclient::models::V1CubeMeta;
-use datafusion::{
-    arrow::datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit},
-    datasource::{self, TableProvider},
-    error::DataFusionError,
-    execution::context::SessionState as DFSessionState,
-    logical_plan::Expr,
-    physical_plan::{udaf::AggregateUDF, udf::ScalarUDF, udtf::TableUDF, ExecutionPlan},
-    sql::planner::ContextProvider,
-};
-
-use crate::{
-    compile::MetaContext,
-    sql::{
-        session::DatabaseProtocol, temp_tables::TempTableProvider, ColumnType, SessionManager,
-        SessionState,
-    },
-    transport::V1CubeMetaExt,
-    CubeError,
-};
-
-use super::information_schema::mysql::{
-    collations::InfoSchemaCollationsProvider as MySqlSchemaCollationsProvider,
-    columns::InfoSchemaColumnsProvider as MySqlSchemaColumnsProvider,
-    key_column_usage::InfoSchemaKeyColumnUsageProvider as MySqlSchemaKeyColumnUsageProvider,
-    processlist::InfoSchemaProcesslistProvider as MySqlSchemaProcesslistProvider,
-    referential_constraints::InfoSchemaReferentialConstraintsProvider as MySqlSchemaReferentialConstraintsProvider,
-    schemata::InfoSchemaSchemataProvider as MySqlSchemaSchemataProvider,
-    statistics::InfoSchemaStatisticsProvider as MySqlSchemaStatisticsProvider,
-    tables::InfoSchemaTableProvider as MySqlSchemaTableProvider,
-    variables::PerfSchemaVariablesProvider as MySqlPerfSchemaVariablesProvider,
-};
+use datafusion::datasource::TableProvider;
 
 use super::information_schema::postgres::{
     character_sets::InfoSchemaCharacterSetsProvider as PostgresSchemaCharacterSetsProvider,
@@ -58,6 +26,14 @@ use super::information_schema::postgres::{
     PgCatalogTypeProvider, PgCatalogUserProvider, PgCatalogViewsProvider,
     PgPreparedStatementsProvider,
 };
+use crate::{
+    compile::{
+        engine::{CubeContext, CubeTableProvider, TableName},
+        DatabaseProtocol,
+    },
+    sql::temp_tables::TempTableProvider,
+    CubeError,
+};
 
 use super::information_schema::redshift::{
     RedshiftLateBindingViewUnpackedTableProvider, RedshiftStlDdltextProvider,
@@ -66,214 +42,10 @@ use super::information_schema::redshift::{
     RedshiftSvvTablesTableProvider,
 };
 
-#[derive(Clone)]
-pub struct CubeContext {
-    /// Internal state for the context (default)
-    pub state: Arc<DFSessionState>,
-    /// References
-    pub meta: Arc<MetaContext>,
-    pub sessions: Arc<SessionManager>,
-    pub session_state: Arc<SessionState>,
-}
-
-impl CubeContext {
-    pub fn new(
-        state: Arc<DFSessionState>,
-        meta: Arc<MetaContext>,
-        sessions: Arc<SessionManager>,
-        session_state: Arc<SessionState>,
-    ) -> Self {
-        Self {
-            state,
-            meta,
-            sessions,
-            session_state,
-        }
-    }
-
-    pub fn table_name_by_table_provider(
-        &self,
-        table_provider: Arc<dyn datasource::TableProvider>,
-    ) -> Result<String, CubeError> {
-        self.session_state
-            .protocol
-            .table_name_by_table_provider(table_provider)
-    }
-
-    pub fn get_function<T>(&self, name: &str, udfs: &HashMap<String, Arc<T>>) -> Option<Arc<T>> {
-        if name.starts_with("pg_catalog.") {
-            return udfs.get(&format!("{}", &name[11..name.len()])).cloned();
-        }
-
-        udfs.get(name).cloned()
-    }
-}
-
-impl ContextProvider for CubeContext {
-    fn get_table_provider(
-        &self,
-        tr: datafusion::catalog::TableReference,
-    ) -> Option<std::sync::Arc<dyn datasource::TableProvider>> {
-        return self.session_state.protocol.get_provider(&self.clone(), tr);
-    }
-
-    fn get_function_meta(&self, name: &str) -> Option<Arc<ScalarUDF>> {
-        self.get_function(name, &self.state.scalar_functions)
-    }
-
-    fn get_aggregate_meta(&self, name: &str) -> Option<Arc<AggregateUDF>> {
-        self.get_function(name, &self.state.aggregate_functions)
-    }
-
-    fn get_table_function_meta(&self, name: &str) -> Option<Arc<TableUDF>> {
-        self.get_function(name, &self.state.table_functions)
-    }
-
-    fn get_variable_type(&self, _variable_names: &[String]) -> Option<DataType> {
-        Some(DataType::Utf8)
-    }
-}
-
 impl DatabaseProtocol {
-    fn get_provider(
-        &self,
-        context: &CubeContext,
-        tr: datafusion::catalog::TableReference,
-    ) -> Option<std::sync::Arc<dyn datasource::TableProvider>> {
-        match self {
-            DatabaseProtocol::MySQL => self.get_mysql_provider(context, tr),
-            DatabaseProtocol::PostgreSQL => self.get_postgres_provider(context, tr),
-        }
-    }
-
-    pub fn table_name_by_table_provider(
-        &self,
-        table_provider: Arc<dyn datasource::TableProvider>,
-    ) -> Result<String, CubeError> {
-        match self {
-            DatabaseProtocol::MySQL => self.get_mysql_table_name(table_provider),
-            DatabaseProtocol::PostgreSQL => self.get_postgres_table_name(table_provider),
-        }
-    }
-
-    pub fn get_mysql_table_name(
-        &self,
-        table_provider: Arc<dyn datasource::TableProvider>,
-    ) -> Result<String, CubeError> {
-        let any = table_provider.as_any();
-        Ok(if let Some(t) = any.downcast_ref::<CubeTableProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaTableProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaColumnsProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaStatisticsProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaKeyColumnUsageProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaSchemataProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaReferentialConstraintsProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaCollationsProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlPerfSchemaVariablesProvider>() {
-            t.table_name().to_string()
-        } else if let Some(t) = any.downcast_ref::<MySqlSchemaProcesslistProvider>() {
-            t.table_name().to_string()
-        } else {
-            return Err(CubeError::internal(format!(
-                "Unknown table provider with schema: {:?}",
-                table_provider.schema()
-            )));
-        })
-    }
-
-    fn get_mysql_provider(
-        &self,
-        context: &CubeContext,
-        tr: datafusion::catalog::TableReference,
-    ) -> Option<std::sync::Arc<dyn datasource::TableProvider>> {
-        let (db, table) = match tr {
-            datafusion::catalog::TableReference::Partial { schema, table, .. } => {
-                (schema.to_ascii_lowercase(), table.to_ascii_lowercase())
-            }
-            datafusion::catalog::TableReference::Full {
-                catalog: _,
-                schema,
-                table,
-            } => (schema.to_ascii_lowercase(), table.to_ascii_lowercase()),
-            datafusion::catalog::TableReference::Bare { table } => {
-                ("db".to_string(), table.to_ascii_lowercase())
-            }
-        };
-
-        match db.as_str() {
-            "db" => {
-                if let Some(cube) = context
-                    .meta
-                    .cubes
-                    .iter()
-                    .find(|c| c.name.eq_ignore_ascii_case(&table))
-                {
-                    // TODO .clone()
-                    return Some(Arc::new(CubeTableProvider::new(cube.clone())));
-                } else {
-                    return None;
-                }
-            }
-            "information_schema" => match table.as_str() {
-                "tables" => {
-                    return Some(Arc::new(MySqlSchemaTableProvider::new(
-                        context.meta.clone(),
-                    )))
-                }
-                "columns" => {
-                    return Some(Arc::new(MySqlSchemaColumnsProvider::new(
-                        context.meta.clone(),
-                    )))
-                }
-                "statistics" => return Some(Arc::new(MySqlSchemaStatisticsProvider::new())),
-                "key_column_usage" => {
-                    return Some(Arc::new(MySqlSchemaKeyColumnUsageProvider::new()))
-                }
-                "schemata" => return Some(Arc::new(MySqlSchemaSchemataProvider::new())),
-                "processlist" => {
-                    return Some(Arc::new(MySqlSchemaProcesslistProvider::new(
-                        context.sessions.clone(),
-                    )))
-                }
-                "referential_constraints" => {
-                    return Some(Arc::new(MySqlSchemaReferentialConstraintsProvider::new()))
-                }
-                "collations" => return Some(Arc::new(MySqlSchemaCollationsProvider::new())),
-                _ => return None,
-            },
-            "performance_schema" => match table.as_str() {
-                "global_variables" => {
-                    return Some(Arc::new(MySqlPerfSchemaVariablesProvider::new(
-                        "performance_schema.global_variables".to_string(),
-                        context
-                            .sessions
-                            .server
-                            .all_variables(context.session_state.protocol.clone()),
-                    )))
-                }
-                "session_variables" => {
-                    return Some(Arc::new(MySqlPerfSchemaVariablesProvider::new(
-                        "performance_schema.session_variables".to_string(),
-                        context.session_state.all_variables(),
-                    )))
-                }
-                _ => return None,
-            },
-            _ => return None,
-        }
-    }
-
     pub fn get_postgres_table_name(
         &self,
-        table_provider: Arc<dyn datasource::TableProvider>,
+        table_provider: Arc<dyn TableProvider>,
     ) -> Result<String, CubeError> {
         let any = table_provider.as_any();
         Ok(if let Some(t) = any.downcast_ref::<TempTableProvider>() {
@@ -386,11 +158,11 @@ impl DatabaseProtocol {
         })
     }
 
-    fn get_postgres_provider(
+    pub(crate) fn get_postgres_provider(
         &self,
         context: &CubeContext,
         tr: datafusion::catalog::TableReference,
-    ) -> Option<std::sync::Arc<dyn datasource::TableProvider>> {
+    ) -> Option<Arc<dyn TableProvider>> {
         let (_, schema, table) = match tr {
             datafusion::catalog::TableReference::Partial { schema, table, .. } => (
                 context.session_state.database().unwrap_or("db".to_string()),
@@ -612,82 +384,5 @@ impl DatabaseProtocol {
         }
 
         None
-    }
-}
-
-pub trait TableName {
-    fn table_name(&self) -> &str;
-}
-
-pub struct CubeTableProvider {
-    cube: V1CubeMeta,
-}
-
-impl CubeTableProvider {
-    pub fn new(cube: V1CubeMeta) -> Self {
-        Self { cube }
-    }
-}
-
-impl TableName for CubeTableProvider {
-    fn table_name(&self) -> &str {
-        &self.cube.name
-    }
-}
-
-#[async_trait]
-impl TableProvider for CubeTableProvider {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn schema(&self) -> SchemaRef {
-        Arc::new(Schema::new(
-            self.cube
-                .get_columns()
-                .into_iter()
-                .map(|c| {
-                    Field::new(
-                        c.get_name(),
-                        match c.get_column_type() {
-                            ColumnType::Date(large) => {
-                                if large {
-                                    DataType::Date64
-                                } else {
-                                    DataType::Date32
-                                }
-                            }
-                            ColumnType::Interval(unit) => DataType::Interval(unit),
-                            ColumnType::String => DataType::Utf8,
-                            ColumnType::VarStr => DataType::Utf8,
-                            ColumnType::Boolean => DataType::Boolean,
-                            ColumnType::Double => DataType::Float64,
-                            ColumnType::Int8 => DataType::Int64,
-                            ColumnType::Int32 => DataType::Int64,
-                            ColumnType::Int64 => DataType::Int64,
-                            ColumnType::Blob => DataType::Utf8,
-                            ColumnType::Decimal(p, s) => DataType::Decimal(p, s),
-                            ColumnType::List(field) => DataType::List(field.clone()),
-                            ColumnType::Timestamp => {
-                                DataType::Timestamp(TimeUnit::Millisecond, None)
-                            }
-                        },
-                        true,
-                    )
-                })
-                .collect(),
-        ))
-    }
-
-    async fn scan(
-        &self,
-        _projection: &Option<Vec<usize>>,
-        _filters: &[Expr],
-        _limit: Option<usize>,
-    ) -> Result<Arc<dyn ExecutionPlan>, DataFusionError> {
-        Err(DataFusionError::Plan(format!(
-            "Not rewritten table scan node for '{}' cube",
-            self.cube.name
-        )))
     }
 }
