+++ rust/cubesql/cubesql/src/compile/engine/information_schema/mysql/key_column_usage.rs
@@ -17,7 +17,7 @@ use super::utils::{
     new_boolean_array_with_placeholder, new_string_array_with_placeholder,
     new_uint32_array_with_placeholder,
 };
-use crate::compile::engine::provider::TableName;
+use crate::compile::engine::context::TableName;
 
 pub struct InfoSchemaKeyColumnUsageProvider {}
