+++ rust/cubesql/cubesql/src/compile/engine/information_schema/mysql/referential_constraints.rs
@@ -14,7 +14,7 @@ use datafusion::{
 };
 
 use super::utils::new_string_array_with_placeholder;
-use crate::compile::engine::provider::TableName;
+use crate::compile::engine::context::TableName;
 
 pub struct InfoSchemaReferentialConstraintsProvider {}
