+++ rust/cubesql/cubesql/src/compile/engine/information_schema/mysql/tables.rs
@@ -1,6 +1,6 @@
 use std::{any::Any, sync::Arc};
 
-use crate::{compile::engine::provider::TableName, transport::MetaContext};
+use crate::{compile::engine::context::TableName, transport::MetaContext};
 use async_trait::async_trait;
 use datafusion::{
     arrow::{
