+++ rust/cubesql/cubesql/src/compile/engine/mod.rs
@@ -1,5 +1,12 @@
-pub mod context;
 pub mod df;
 pub mod information_schema;
-pub mod provider;
 pub mod udf;
+
+mod context;
+mod context_mysql;
+mod context_postgresql;
+mod variable_provider;
+
+// Public API
+pub use context::*;
+pub use variable_provider::*;
